[Move #234] modal 폴더 생성 후 modal 관련 컴포넌트 분리
