+++ components/modal/GroupMafiaModal.tsx
